requirements.txt
python-dotenv>=1.0.0
hmmlearn>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
streamlit>=1.30.0
plotly>=5.18.0
alpaca-py>=0.20.0
